weather_data_service/requirements.txt
Django==3.0
djangorestframework==3.10.3
djangogrpcframework==0.2.1
grpcio==1.66.1
grpcio-tools==1.66.1
gunicorn==20.1.0
mysqlclient==2.0.3
django-cors-headers==3.7.0
python-decouple==3.4
django-filter==21.1
djangorestframework-simplejwt==4.7.2
drf-yasg==1.20.0